requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.4
easydict==1.13
filterpy==1.4.5
flatbuffers==25.9.23
fonttools==4.60.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
insightface==0.7.3
joblib==1.5.2
kiwisolver==1.4.9
lap==0.5.12
lazy_loader==0.4
matplotlib==3.10.6
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onnx==1.19.0
onnxruntime-gpu==1.22.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
prettytable==3.16.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
PyQt6==6.9.1
PyQt6-Qt6==6.9.2
PyQt6_sip==13.10.2
pytest==8.4.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
simsimd==6.5.3
six==1.17.0
sort-tracker-py==1.0.2
sqlite-vec==0.1.6
stringzilla==4.0.14
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.9.20
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
